% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{cluster.hierarchical}
\alias{cluster.hierarchical}
\title{Bottom-up Hierarchical Clustering using AGNES algorithm, 
including single-linkage, complete-linkage and average-linkage method}
\usage{
cluster.hierarchical(x, K = 3, method = "min")
}
\arguments{
\item{x}{An array of multiple vectors, which different rows are different vectors}

\item{K}{Cluster numbers}

\item{method}{the method using to calculate the distance between two clusters. 
('min', 'max', 'mean', default: 'min')}
}
\value{
A vector of the clusters id for each input vectors
}
\description{
Bottom-up Hierarchical Clustering using AGNES algorithm, 
including single-linkage, complete-linkage and average-linkage method
}
\examples{
\dontrun{
x <- iris[1:100,1:4]
label = cluster.prototype(x, K=2,method='mean')
}
}
