% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{cluster.prototype}
\alias{cluster.prototype}
\title{K-means cluster and K-medians cluster for a group of multiple vectors}
\usage{
cluster.prototype(
  x,
  K = 3,
  center = NULL,
  method = "mean",
  eps = 1e-05,
  i_max = 1000
)
}
\arguments{
\item{x}{An array of multiple vectors, which different rows are different vectors}

\item{K}{Cluster numbers}

\item{center}{The initial center coordinates (default: NULL)}

\item{method}{str. "mean" (K-means) or "median" (K-medians) (default: "mean")}

\item{eps}{Conditions for Judging Convergence (default: 1e-5)}

\item{i_max}{Maximum number of iterations (default: 1000)}
}
\value{
A list containing clustering results of different vectors and cluster centers
}
\description{
K-means cluster and K-medians cluster for a group of multiple vectors
}
\examples{
\dontrun{
x <- iris[(iris$Species == 'setosa') | (iris$Species == 'versicolor'),]
result <- cluster.prototype(x[1:4], K=2)
x['cluster'] = result[[1]]
pca_iris = princomp(x[,1:4])$scores[,1:2]
plot(pca_iris, t='n')
text(pca_iris, labels=x$Species,col=x$cluster) 
}

}
