% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{cluster_distanse}
\alias{cluster_distanse}
\title{Calculate the distance between two clusters}
\usage{
cluster_distanse(P, Q, method = "min")
}
\arguments{
\item{P}{An array of multiple vectors, which different rows are different vectors}

\item{Q}{An array of multiple vectors, which different rows are different vectors}

\item{method}{the method using to calculate the distance between two clusters. 
('min', 'max', 'mean', default: 'min')}
}
\value{
The distance between two vector groups
}
\description{
Calculate the distance between two clusters
}
