% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{euclidean}
\alias{euclidean}
\title{Calculate the Euclidean distance between two vectors}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
the Euclidean distance between two array of vectors
}
\description{
Calculate the Euclidean distance between two vectors
}
\examples{
\dontrun{
x <- c(1,2,0)
y <- c(2,1,1)
euclidean(x,y)
}
}
