% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{pairwise_distance}
\alias{pairwise_distance}
\title{Calculate the pairwise Euclidean distance between two vector groups}
\usage{
pairwise_distance(x, y)
}
\arguments{
\item{x}{An array of multiple vectors, which different rows are different vectors}

\item{y}{An array of multiple vectors, which different rows are different vectors}
}
\value{
Adjacency matrix between two sets of vectors
}
\description{
Calculate the pairwise Euclidean distance between two vector groups
}
\examples{
\dontrun{
x <- iris[,1:4][iris$Species == 'setosa',]
y <- iris[,1:4][iris$Species == 'versicolor',]
dis <- pairwise_distance(x,y)
}
}
